+++ R/fit.R
@@ -181,7 +181,19 @@ refit_multiple_optimizers <- function(fit,
 #' @param accept_singular (`flag`)\cr whether singular design matrices are reduced
 #'   to full rank automatically and additional coefficient estimates will be missing.
 #' @param optimizers (`list`)\cr of optimizers created from [h_get_optimizers()].
+#' @param drop_visit_levels (`flag`)\cr whether to drop levels for visit variable, if visit variable is a factor.
 #' @param ... Additional arguments passed to [h_get_optimizers()].
+#'
+#' @details
+#' `drop_visit_levels`:
+#' This flag will decide whether unobserved visits will be kept for analysis.
+#' For example, if you have only VIS1, VIS3 and VIS4, by default
+#' they are treated to be equally spaced, the distance from VIS1 to VIS3, and from VIS3 to VIS4,
+#' are both 1. However, you can manually convert this visit into a factor, with
+#' `levels = c("VIS1", "VIS2", "VIS3", "VIS4")`, and also use `drop_visits_levels = FALSE`,
+#' then the distance from VIS1 to VIS3 will be 2, as "VIS2" is a valid visit.
+#' However, please be cautious because this can lead to convergence issues for unstructured covariance
+#' structure because there are no observations at that missing visit.
 #'
 #' @return List of class `mmrm_control` with the control parameters.
 #' @export
@@ -195,12 +207,14 @@ mmrm_control <- function(n_cores = 1L,
                          method = c("Satterthwaite", "Kenward-Roger", "Kenward-Roger-Linear"),
                          start = NULL,
                          accept_singular = TRUE,
+                         drop_visit_levels = TRUE,
                          ...,
                          optimizers = h_get_optimizers(...)) {
   assert_int(n_cores, lower = 1L)
   assert_character(method)
   assert_numeric(start, null.ok = TRUE)
   assert_flag(accept_singular)
+  assert_flag(drop_visit_levels)
   assert_list(optimizers, names = "unique", types = c("function", "partial"))
   method <- match.arg(method)
   structure(
@@ -209,7 +223,8 @@ mmrm_control <- function(n_cores = 1L,
       start = start,
       accept_singular = accept_singular,
       method = method,
-      n_cores = as.integer(n_cores)
+      n_cores = as.integer(n_cores),
+      drop_visit_levels = drop_visit_levels
     ),
     class = "mmrm_control"
   )